agents/requirements.txt
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
zeroclaw-tools>=0.1.0
